updates for big integers

+++ selection_sort.c
@@ -1,15 +1,15 @@
 #include <stdio.h>
 
-void swap(int *a, int *b)
+void swap(long long int *a, long long int *b)
 {
-	int temp = *a;
+	long long int temp = *a;
 	*a = *b;
 	*b = temp;
 }
 
-void sort(int arr[], int n)
+void sort(long long int arr[], long long int n)
 {
-	int i, j, min;
+	long long int i, j, min;
 	for (i = 0; i < n-1; i++)
 	{
 		min = i;
@@ -20,18 +20,18 @@ void sort(int arr[], int n)
 	}
 }
 
-void printArray(int arr[], int n)
+void printArray(long long int arr[], long long int n)
 {
-	int i;
+	long long int i;
 	for (i=0; i < n; i++)
-		printf("%d ", arr[i]);
+		printf("%lld ", arr[i]);
 	printf("\n");
 }
 
 int main()
 {
-	int arr[] = {215,211,74,217,212};
-	int n = sizeof(arr)/sizeof(arr[0]);
+	long long int arr[] = {215,211,74,217,212};
+	long long int n = sizeof(arr)/sizeof(arr[0]);
 	printf("Array before sorting : ");
 	printArray(arr, n);
 	sort(arr, n);
